requirements.txt
gunicorn==20.0.4
Django==3.1.12
django-cors-headers==3.6.0
django-massmailer==0.6
django-oauth-toolkit==1.3.3
django-environ
django-jsonview==2.0.0
get-docker-secret==1.0.1
mysqlclient==2.0.2
mysql-connector-python
celery==5.0.5
arrow==0.17.0
redis==3.5.3
PyJWT==1.7
PyMySQL==1.0.2